switch to the F5 nginx controller
